// src/interfaces/CharacterInterfaces.ts
export enum SKIN_TYPES {
  L2D = 1,
  IMAGE = 2,
}

export interface SetShowSkinParams {
  index: number;
  type: SKIN_TYPES | null;
}